requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Scrapy
scrapy==2.11.0
scrapy-playwright==0.0.34

# Content Processing
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1
bleach==6.1.0
markdown==3.5.2

# Job Queue (Optional)
redis==5.0.1
rq==1.16.1

# Utilities
python-slugify==8.0.1
python-multipart==0.0.6
httpx==0.26.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==24.1.1
flake8==7.0.0
